:recycle: dispatch extract protocol processing to separate use case
